pyproject.toml
[tool.poetry]
name = "digital_pathology"
version = "1.0.0"
description = "Digital Pathology"
authors = ["Israel Llorens <sanchezis@hotmail.com>"]
license = "MIT OR EUPL-1.2"  # SPDX expression
classifiers = [
    "License :: OSI Approved :: MIT License",
    "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.24"  # scikit-image 0.19.3 breaks with numpy>=1.24
pandas = "^2.0.0"
pyspark = "3.4.4"
pyarrow = "14.0.2"  # Use a newer version with pre-built wheels
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"
boto3 = "^1.37.18"
matplotlib = "3.9.3"
seaborn = "^0.13.2"
tifffile = "2024.9.20"
tiffslide = "^2.5.0"
slideio = "^2.7.0"
plotly = "^6.0.1"
openslide-python = "1.3.1"
openslide-bin = { version = "^4.0.0.6", markers = "sys_platform == 'darwin' or sys_platform == 'linux'" }
scipy = "^1.8.0"
scikit-image = "0.24.0"
scikit-learn = "1.5.2"
shapely = "2.0.6"
tensorflow = {version = "^2.16.1", extras = ["cuda"], markers = "sys_platform != 'darwin'"}
tensorflow-macos = {version = "^2.16.2", markers = "sys_platform == 'darwin'"}
tensorflow-metal = {version = "^1.2.0", markers = "sys_platform == 'darwin'"}
stardist = "^0.9.1"
histomicstk = "^1.4.0"
ipykernel = "^6.29.5"
pillow = "11.0.0"
statsmodels = "^0.14.4"
opencv-python = "4.9.0.80"
tiatoolbox = "^1.6.0"
torch-geometric = "^2.6.1"
ujson = "^5.10.0"
llvmlite = "0.42.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
black = "^23.10.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"